Inline hot functions used when choosing min column
**Description**
Add the inline attribute to make it more likely that these small
functions would be inlined.

**Motivation**
This change is important because these functions are called very
often when choosing the next column to eliminate in the solver.

The initial profile had ~70% of time spent in the `choose_column`
function, and ~30% in each of `Column::index` and `Column::size`.

Here are the benchmark results when compiling the regression test
using the `release` profile:

| Command        | Mean [s]       | Min [s] | Max [s] | Relative    |
|:---------------|---------------:|--------:|--------:|------------:|
| `./inlined`    | 11.202 ± 0.042 | 11.161  | 11.289  | 1.00        |
| `./noninlined` | 18.529 ± 0.323 | 17.781  | 18.882  | 1.65 ± 0.03 |