Polyomino: Change PShape representation
This commit introduces changes produced in the following discussion:
https://github.com/declanvk/dancing-links/pull/10.
It changes inner representation of PShape (from Vec<Vec<_>> to simple Vec<_>),
treating it as a struct with attached impl block, rather than simple
type alias.
It also moves constraints imposed on that type to `new` function of that
type (type constructor) - earlier assertion of these constraints was
done by `Polyomino` struct, which should not be its responsibility.